Rewrite teapot.accept and fix a discovered bug in routing
This was the most weird bug odyssey for some time now. It all started
with LanguagePreferenceList not working correctly due to
set/dictionary-ordering-dependent code. After rewriting the
teapot.accept library (which was due, as it was legacy code from PyXWF),
it turns out that this very dependency on dict ordering has covered
another bug in the routing, where some code *also* was dependent on dict
ordering -- either in the same or in a sufficiently similar way to be
undiscovered for a looong time.